MAINT: Move np.char methods to np.strings
