+++ numpy/_core/defchararray.py
@@ -18,13 +18,11 @@ The preferred alias for `defchararray` is `numpy.char`.
 import functools
 
 from .._utils import set_module
-from .numerictypes import (
-    bytes_, str_, integer, int_, object_, character)
-from .numeric import ndarray, array as narray
-from numpy._core.multiarray import _vec_string, compare_chararrays
+from .numerictypes import bytes_, str_, character
+from .numeric import ndarray, array as narray, asarray as asnarray
+from numpy._core.multiarray import compare_chararrays
 from numpy._core import overrides
-from numpy._utils import asbytes
-import numpy
+from numpy.strings import *
 
 __all__ = [
     'equal', 'not_equal', 'greater_equal', 'less_equal',
@@ -39,64 +37,8 @@ __all__ = [
     ]
 
 
-_globalvar = 0
-
 array_function_dispatch = functools.partial(
     overrides.array_function_dispatch, module='numpy.char')
-
-
-def _is_unicode(arr):
-    """Returns True if arr is a string or a string array with a dtype that
-    represents a unicode string, otherwise returns False.
-
-    """
-    if (isinstance(arr, str) or
-            issubclass(numpy.asarray(arr).dtype.type, str)):
-        return True
-    return False
-
-
-def _to_bytes_or_str_array(result, output_dtype_like):
-    """
-    Helper function to cast a result back into an array
-    with the appropriate dtype if an object array must be used
-    as an intermediary.
-    """
-    output_dtype_like = numpy.asarray(output_dtype_like)
-    if result.size == 0:
-        # Calling asarray & tolist in an empty array would result
-        # in losing shape information
-        return result.astype(output_dtype_like.dtype)
-    ret = numpy.asarray(result.tolist())
-    return ret.astype(type(output_dtype_like.dtype)(_get_num_chars(ret)))
-
-
-def _clean_args(*args):
-    """
-    Helper function for delegating arguments to Python string
-    functions.
-
-    Many of the Python string operations that have optional arguments
-    do not use 'None' to indicate a default value.  In these cases,
-    we need to remove all None arguments, and those following them.
-    """
-    newargs = []
-    for chk in args:
-        if chk is None:
-            break
-        newargs.append(chk)
-    return newargs
-
-
-def _get_num_chars(a):
-    """
-    Helper function that returns the number of characters per field in
-    a string or unicode array.  This is to abstract out the fact that
-    for a unicode array this is itemsize / 4.
-    """
-    if issubclass(a.dtype.type, str_):
-        return a.itemsize // 4
-    return a.itemsize
 
 
 def _binary_op_dispatcher(x1, x2):
@@ -300,1798 +242,6 @@ def less(x1, x2):
     
     """
     return compare_chararrays(x1, x2, '<', True)
-
-
-def _unary_op_dispatcher(a):
-    return (a,)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def str_len(a):
-    """
-    Return len(a) element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of integers
-
-    See Also
-    --------
-    len
-
-    Examples
-    --------
-    >>> a = np.array(['Grace Hopper Conference', 'Open Source Day'])
-    >>> np.char.str_len(a)
-    array([23, 15])
-    >>> a = np.array(['\u0420', '\u043e'])
-    >>> np.char.str_len(a)
-    array([1, 1])
-    >>> a = np.array([['hello', 'world'], ['\u0420', '\u043e']])
-    >>> np.char.str_len(a)
-    array([[5, 5], [1, 1]])
-    
-    """
-    return numpy._core.umath.str_len(a)
-
-
-@array_function_dispatch(_binary_op_dispatcher)
-def add(x1, x2):
-    """
-    Return element-wise string concatenation for two arrays of str or unicode.
-
-    Arrays `x1` and `x2` must have the same shape.
-
-    Parameters
-    ----------
-    x1 : array_like of str or unicode
-        Input array.
-    x2 : array_like of str or unicode
-        Input array.
-
-    Returns
-    -------
-    add : ndarray
-        Output array of `bytes_` or `str_`, depending on input types
-        of the same shape as `x1` and `x2`.
-
-    Examples
-    --------
-    >>> x = np.array(['1'])
-    >>> y = np.array(['a'])
-    >>> np.char.add(x, y)
-    array(['1a'], dtype='<U2')
-
-    >>> x = np.array(['1', '2'])
-    >>> y = np.array(['a', 'b'])
-    >>> np.char.add(x, y)
-    array(['1a', '2b'], dtype='<U2')
-
-    """
-    arr1 = numpy.asarray(x1)
-    arr2 = numpy.asarray(x2)
-
-    if type(arr1.dtype) != type(arr2.dtype):
-        # Enforce this for now.  The solution to it will be implement add
-        # as a ufunc.  It never worked right on Python 3: bytes + unicode gave
-        # nonsense unicode + bytes errored, and unicode + object used the
-        # object dtype itemsize as num chars (worked on short strings).
-        # bytes + void worked but promoting void->bytes is dubious also.
-        raise TypeError(
-            "np.char.add() requires both arrays of the same dtype kind, but "
-            f"got dtypes: '{arr1.dtype}' and '{arr2.dtype}' (the few cases "
-            "where this used to work often lead to incorrect results).")
-
-    return numpy.add(x1, x2)
-
-def _multiply_dispatcher(a, i):
-    return (a,)
-
-
-@array_function_dispatch(_multiply_dispatcher)
-def multiply(a, i):
-    """
-    Return (a * i), that is string multiple concatenation,
-    element-wise.
-
-    Values in `i` of less than 0 are treated as 0 (which yields an
-    empty string).
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    i : array_like of ints
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input types
-    
-    Examples
-    --------
-    >>> a = np.array(["a", "b", "c"])
-    >>> np.char.multiply(a, 3)
-    array(['aaa', 'bbb', 'ccc'], dtype='<U3')
-    >>> i = np.array([1, 2, 3])
-    >>> np.char.multiply(a, i)
-    array(['a', 'bb', 'ccc'], dtype='<U3')
-    >>> np.char.multiply(np.array(['a']), i)
-    array(['a', 'aa', 'aaa'], dtype='<U3')
-    >>> a = np.array(['a', 'b', 'c', 'd', 'e', 'f']).reshape((2, 3))
-    >>> np.char.multiply(a, 3)
-    array([['aaa', 'bbb', 'ccc'],
-           ['ddd', 'eee', 'fff']], dtype='<U3')
-    >>> np.char.multiply(a, i)
-    array([['a', 'bb', 'ccc'],
-           ['d', 'ee', 'fff']], dtype='<U3')
-           
-    """
-    a_arr = numpy.asarray(a)
-    i_arr = numpy.asarray(i)
-    if not issubclass(i_arr.dtype.type, integer):
-        raise ValueError("Can only multiply by integers")
-    out_size = _get_num_chars(a_arr) * max(int(i_arr.max()), 0)
-    return _vec_string(
-        a_arr, type(a_arr.dtype)(out_size), '__mul__', (i_arr,))
-
-
-def _mod_dispatcher(a, values):
-    return (a, values)
-
-
-@array_function_dispatch(_mod_dispatcher)
-def mod(a, values):
-    """
-    Return (a % i), that is pre-Python 2.6 string formatting
-    (interpolation), element-wise for a pair of array_likes of str
-    or unicode.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    values : array_like of values
-       These values will be element-wise interpolated into the string.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input types
-        
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, '__mod__', (values,)), a)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def capitalize(a):
-    """
-    Return a copy of `a` with only the first character of each element
-    capitalized.
-
-    Calls :meth:`str.capitalize` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-        Input array of strings to capitalize.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input
-        types
-
-    See Also
-    --------
-    str.capitalize
-
-    Examples
-    --------
-    >>> c = np.array(['a1b2','1b2a','b2a1','2a1b'],'S4'); c
-    array(['a1b2', '1b2a', 'b2a1', '2a1b'],
-        dtype='|S4')
-    >>> np.char.capitalize(c)
-    array(['A1b2', '1b2a', 'B2a1', '2a1b'],
-        dtype='|S4')
-
-    """
-    a_arr = numpy.asarray(a)
-    return _vec_string(a_arr, a_arr.dtype, 'capitalize')
-
-
-def _center_dispatcher(a, width, fillchar=None):
-    return (a,)
-
-
-@array_function_dispatch(_center_dispatcher)
-def center(a, width, fillchar=' '):
-    """
-    Return a copy of `a` with its elements centered in a string of
-    length `width`.
-
-    Calls :meth:`str.center` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    width : int
-        The length of the resulting strings
-    fillchar : str or unicode, optional
-        The padding character to use (default is space).
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input
-        types
-
-    See Also
-    --------
-    str.center
-    
-    Notes
-    -----
-    This function is intended to work with arrays of strings.  The
-    fill character is not applied to numeric types.
-
-    Examples
-    --------
-    >>> c = np.array(['a1b2','1b2a','b2a1','2a1b']); c
-    array(['a1b2', '1b2a', 'b2a1', '2a1b'], dtype='<U4')
-    >>> np.char.center(c, width=9)
-    array(['   a1b2  ', '   1b2a  ', '   b2a1  ', '   2a1b  '], dtype='<U9')
-    >>> np.char.center(c, width=9, fillchar='*')
-    array(['***a1b2**', '***1b2a**', '***b2a1**', '***2a1b**'], dtype='<U9')
-    >>> np.char.center(c, width=1)
-    array(['a', '1', 'b', '2'], dtype='<U1')
-
-    """
-    a_arr = numpy.asarray(a)
-    width_arr = numpy.asarray(width)
-    size = int(numpy.max(width_arr.flat))
-    if numpy.issubdtype(a_arr.dtype, numpy.bytes_):
-        fillchar = asbytes(fillchar)
-    return _vec_string(
-        a_arr, type(a_arr.dtype)(size), 'center', (width_arr, fillchar))
-
-
-def _count_dispatcher(a, sub, start=None, end=None):
-    return (a,)
-
-
-@array_function_dispatch(_count_dispatcher)
-def count(a, sub, start=0, end=None):
-    """
-    Returns an array with the number of non-overlapping occurrences of
-    substring `sub` in the range [`start`, `end`].
-
-    Calls :meth:`str.count` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    sub : str or unicode
-       The substring to search for.
-
-    start, end : int, optional
-       Optional arguments `start` and `end` are interpreted as slice
-       notation to specify the range in which to count.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of ints.
-
-    See Also
-    --------
-    str.count
-
-    Examples
-    --------
-    >>> c = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> c
-    array(['aAaAaA', '  aA  ', 'abBABba'], dtype='<U7')
-    >>> np.char.count(c, 'A')
-    array([3, 1, 1])
-    >>> np.char.count(c, 'aA')
-    array([3, 1, 0])
-    >>> np.char.count(c, 'A', start=1, end=4)
-    array([2, 1, 1])
-    >>> np.char.count(c, 'A', start=1, end=3)
-    array([1, 0, 0])
-
-    """
-    end = end if end is not None else numpy.iinfo(numpy.int64).max
-    return numpy._core.umath.count(a, sub, start, end)
-
-
-def _code_dispatcher(a, encoding=None, errors=None):
-    return (a,)
-
-
-@array_function_dispatch(_code_dispatcher)
-def decode(a, encoding=None, errors=None):
-    r"""
-    Calls :meth:`bytes.decode` element-wise.
-
-    The set of available codecs comes from the Python standard library,
-    and may be extended at runtime.  For more information, see the
-    :mod:`codecs` module.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    encoding : str, optional
-       The name of an encoding
-
-    errors : str, optional
-       Specifies how to handle encoding errors
-
-    Returns
-    -------
-    out : ndarray
-
-    See Also
-    --------
-    :py:meth:`bytes.decode`
-
-    Notes
-    -----
-    The type of the result will depend on the encoding specified.
-
-    Examples
-    --------
-    >>> c = np.array([b'\x81\xc1\x81\xc1\x81\xc1', b'@@\x81\xc1@@',
-    ...               b'\x81\x82\xc2\xc1\xc2\x82\x81'])
-    >>> c
-    array([b'\x81\xc1\x81\xc1\x81\xc1', b'@@\x81\xc1@@',
-           b'\x81\x82\xc2\xc1\xc2\x82\x81'], dtype='|S7')
-    >>> np.char.decode(c, encoding='cp037')
-    array(['aAaAaA', '  aA  ', 'abBABba'], dtype='<U7')
-
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, 'decode', _clean_args(encoding, errors)),
-        numpy.str_(''))
-
-
-@array_function_dispatch(_code_dispatcher)
-def encode(a, encoding=None, errors=None):
-    """
-    Calls :meth:`str.encode` element-wise.
-
-    The set of available codecs comes from the Python standard library,
-    and may be extended at runtime. For more information, see the
-    :mod:`codecs` module.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    encoding : str, optional
-       The name of an encoding
-
-    errors : str, optional
-       Specifies how to handle encoding errors
-
-    Returns
-    -------
-    out : ndarray
-
-    See Also
-    --------
-    str.encode
-
-    Notes
-    -----
-    The type of the result will depend on the encoding specified.
-
-    Examples
-    --------
-    >>> a = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> np.char.encode(a, encoding='cp037')
-    array([b'\x81\xc1\x81\xc1\x81\xc1', b'@@\x81\xc1@@',
-       b'\x81\x82\xc2\xc1\xc2\x82\x81'], dtype='|S7')
-       
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, 'encode', _clean_args(encoding, errors)),
-        numpy.bytes_(b''))
-
-def _endswith_dispatcher(a, suffix, start=None, end=None):
-    return (a,)
-
-
-@array_function_dispatch(_endswith_dispatcher)
-def endswith(a, suffix, start=0, end=None):
-    """
-    Returns a boolean array which is `True` where the string element
-    in `a` ends with `suffix`, otherwise `False`.
-
-    Calls :meth:`str.endswith` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    suffix : str
-
-    start, end : int, optional
-        With optional `start`, test beginning at that position. With
-        optional `end`, stop comparing at that position.
-
-    Returns
-    -------
-    out : ndarray
-        Outputs an array of bools.
-
-    See Also
-    --------
-    str.endswith
-
-    Examples
-    --------
-    >>> s = np.array(['foo', 'bar'])
-    >>> s[0] = 'foo'
-    >>> s[1] = 'bar'
-    >>> s
-    array(['foo', 'bar'], dtype='<U3')
-    >>> np.char.endswith(s, 'ar')
-    array([False,  True])
-    >>> np.char.endswith(s, 'a', start=1, end=2)
-    array([False,  True])
-
-    """
-    end = end if end is not None else numpy.iinfo(numpy.int_).max
-    return numpy._core.umath.endswith(a, suffix, start, end)
-
-
-def _expandtabs_dispatcher(a, tabsize=None):
-    return (a,)
-
-
-@array_function_dispatch(_expandtabs_dispatcher)
-def expandtabs(a, tabsize=8):
-    """
-    Return a copy of each string element where all tab characters are
-    replaced by one or more spaces.
-
-    Calls :meth:`str.expandtabs` element-wise.
-
-    Return a copy of each string element where all tab characters are
-    replaced by one or more spaces, depending on the current column
-    and the given `tabsize`. The column number is reset to zero after
-    each newline occurring in the string. This doesn't understand other
-    non-printing characters or escape sequences.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-        Input array
-    tabsize : int, optional
-        Replace tabs with `tabsize` number of spaces.  If not given defaults
-        to 8 spaces.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.expandtabs
-
-    Examples
-    --------
-    >>> a = np.array(['\t\tHello\tworld'])  
-    >>> np.char.expandtabs(a, tabsize=4)  # doctest: +SKIP
-    array(['        Hello   world'], dtype='<U21')  # doctest: +SKIP
-
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, 'expandtabs', (tabsize,)), a)
-
-
-@array_function_dispatch(_count_dispatcher)
-def find(a, sub, start=0, end=None):
-    """
-    For each element, return the lowest index in the string where
-    substring `sub` is found.
-
-    Calls :meth:`str.find` element-wise.
-
-    For each element, return the lowest index in the string where
-    substring `sub` is found, such that `sub` is contained in the
-    range [`start`, `end`].
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    sub : str or unicode
-
-    start, end : int, optional
-        Optional arguments `start` and `end` are interpreted as in
-        slice notation.
-
-    Returns
-    -------
-    out : ndarray or int
-        Output array of ints.  Returns -1 if `sub` is not found.
-
-    See Also
-    --------
-    str.find
-
-    Examples
-    --------
-    >>> a = np.array(["NumPy is a Python library"])
-    >>> np.char.find(a, "Python", start=0, end=None)
-    array([11])
-
-    """
-    end = end if end is not None else numpy.iinfo(numpy.int64).max
-    return numpy._core.umath.find(a, sub, start, end)
-
-@array_function_dispatch(_count_dispatcher)
-def index(a, sub, start=0, end=None):
-    """
-    Like `find`, but raises :exc:`ValueError` when the substring is not found.
-
-    Calls :meth:`str.index` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    sub : str or unicode
-
-    start, end : int, optional
-
-    Returns
-    -------
-    out : ndarray
-        Output array of ints.  Returns -1 if `sub` is not found.
-
-    See Also
-    --------
-    find, str.find
-
-    Examples
-    --------
-    >>> a = np.array(["Computer Science"])
-    >>> np.char.index(a, "Science", start=0, end=None)
-    array([9])
-
-    """
-    return _vec_string(
-        a, int_, 'index', [sub, start] + _clean_args(end))
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isalnum(a):
-    """
-    Returns true for each element if all characters in the string are
-    alphanumeric and there is at least one character, false otherwise.
-
-    Calls :meth:`str.isalnum` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.isalnum
-
-    Examples
-    --------
-    >>> a = np.array(['a', '1', 'a1', '(', ''])
-    >>> np.char.isalnum(a)
-    array([ True,  True,  True, False, False])
-    
-    """
-    return _vec_string(a, numpy.bool, 'isalnum')
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isalpha(a):
-    """
-    Returns true for each element if all characters in the string are
-    alphabetic and there is at least one character, false otherwise.
-
-    Calls :meth:`str.isalpha` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.isalpha
-
-    Examples
-    --------
-    >>> a = np.array(['a', '1', 'a1', ''])
-    >>> np.char.isalpha(a)
-    array([ True,  False,  False,  False])
-
-    """
-    return numpy._core.umath.isalpha(a)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isdigit(a):
-    """
-    Returns true for each element if all characters in the string are
-    digits and there is at least one character, false otherwise.
-
-    Calls :meth:`str.isdigit` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.isdigit
-
-    Examples
-    --------
-    >>> a = np.array(['a', 'b', '0'])
-    >>> np.char.isdigit(a)
-    array([False, False,  True])
-    >>> a = np.array([['a', 'b', '0'], ['c', '1', '2']])
-    >>> np.char.isdigit(a)
-    array([[False, False,  True], [False,  True,  True]])
-    
-    """
-    return numpy._core.umath.isdigit(a)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def islower(a):
-    """
-    Returns true for each element if all cased characters in the
-    string are lowercase and there is at least one cased character,
-    false otherwise.
-
-    Calls :meth:`str.islower` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.islower
-
-    Examples
-    --------
-    >>> np.char.islower("GHC")
-    array(False)
-    >>> np.char.islower("ghc")
-    array(True)
-
-    """
-    return _vec_string(a, numpy.bool, 'islower')
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isspace(a):
-    """
-    Returns true for each element if there are only whitespace
-    characters in the string and there is at least one character,
-    false otherwise.
-
-    Calls :meth:`str.isspace` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.isspace
-
-    Examples
-    --------
-    >>> np.char.isspace(["GHC", "GHC "])
-    array([False, False])
-
-    >>> a = np.array([' ', '\t'])
-    >>> np.char.isspace(a)
-    array([ True,  True])
-    
-    """
-    return numpy._core.umath.isspace(a)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def istitle(a):
-    """
-    Returns true for each element if the element is a titlecased
-    string and there is at least one character, false otherwise.
-
-    Call :meth:`str.istitle` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.istitle
-
-    Examples
-    --------
-    >>> np.char.istitle("Numpy Is Great")
-    array(True)
-
-    >>> np.char.istitle("Numpy is great")
-    array(False)
-    
-    """
-    return _vec_string(a, numpy.bool, 'istitle')
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isupper(a):
-    """
-    Return true for each element if all cased characters in the
-    string are uppercase and there is at least one character, false
-    otherwise.
-
-    Call :meth:`str.isupper` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of bools
-
-    See Also
-    --------
-    str.isupper
-
-    Examples
-    --------
-    >>> np.char.isupper("GHC")
-    array(True)     
-    >>> a = np.array(["hello", "HELLO", "Hello"])
-    >>> np.char.isupper(a)
-    array([False,  True, False]) 
-
-    """
-    return _vec_string(a, numpy.bool, 'isupper')
-
-
-def _join_dispatcher(sep, seq):
-    return (sep, seq)
-
-
-@array_function_dispatch(_join_dispatcher)
-def join(sep, seq):
-    """
-    Return a string which is the concatenation of the strings in the
-    sequence `seq`.
-
-    Calls :meth:`str.join` element-wise.
-
-    Parameters
-    ----------
-    sep : array_like of str or unicode
-    seq : array_like of str or unicode
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input types
-
-    See Also
-    --------
-    str.join
-
-    Examples
-    --------
-    >>> np.char.join('-', 'osd')
-    array('o-s-d', dtype='<U5')
-
-    >>> np.char.join(['-', '.'], ['ghc', 'osd'])
-    array(['g-h-c', 'o.s.d'], dtype='<U5')
-
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(sep, object_, 'join', (seq,)), seq)
-
-
-def _just_dispatcher(a, width, fillchar=None):
-    return (a,)
-
-
-@array_function_dispatch(_just_dispatcher)
-def ljust(a, width, fillchar=' '):
-    """
-    Return an array with the elements of `a` left-justified in a
-    string of length `width`.
-
-    Calls :meth:`str.ljust` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    width : int
-        The length of the resulting strings
-    fillchar : str or unicode, optional
-        The character to use for padding
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.ljust
-
-    Examples
-    --------
-    >>> c = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> np.char.ljust(c, width=3)
-    array(['aAa', '  a', 'abB'], dtype='<U3')
-    
-    """
-    a_arr = numpy.asarray(a)
-    width_arr = numpy.asarray(width)
-    size = int(numpy.max(width_arr.flat))
-    if numpy.issubdtype(a_arr.dtype, numpy.bytes_):
-        fillchar = asbytes(fillchar)
-    return _vec_string(
-        a_arr, type(a_arr.dtype)(size), 'ljust', (width_arr, fillchar))
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def lower(a):
-    """
-    Return an array with the elements converted to lowercase.
-
-    Call :meth:`str.lower` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array.
-
-    Returns
-    -------
-    out : ndarray, {str, unicode}
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.lower
-
-    Examples
-    --------
-    >>> c = np.array(['A1B C', '1BCA', 'BCA1']); c
-    array(['A1B C', '1BCA', 'BCA1'], dtype='<U5')
-    >>> np.char.lower(c)
-    array(['a1b c', '1bca', 'bca1'], dtype='<U5')
-
-    """
-    a_arr = numpy.asarray(a)
-    return _vec_string(a_arr, a_arr.dtype, 'lower')
-
-
-def _strip_dispatcher(a, chars=None):
-    return (a,)
-
-
-@array_function_dispatch(_strip_dispatcher)
-def lstrip(a, chars=None):
-    """
-    For each element in `a`, return a copy with the leading characters
-    removed.
-
-    Parameters
-    ----------
-    a : array_like, with ``bytes_`` or ``unicode_`` dtype
-    chars : scalar with the same dtype as ``a``, optional
-       The `chars` argument is a string specifying the set of
-       characters to be removed. If None, the `chars`
-       argument defaults to removing whitespace. The `chars` argument
-       is not a prefix or suffix; rather, all combinations of its
-       values are stripped.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of ``bytes_`` or ``unicode_`` dtype
-
-    See Also
-    --------
-    str.lstrip
-
-    Examples
-    --------
-    >>> c = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> c
-    array(['aAaAaA', '  aA  ', 'abBABba'], dtype='<U7')
-    # The 'a' variable is unstripped from c[1] because of leading whitespace.
-    >>> np.char.lstrip(c, 'a')
-    array(['AaAaA', '  aA  ', 'bBABba'], dtype='<U7')
-    >>> np.char.lstrip(c, 'A') # leaves c unchanged
-    array(['aAaAaA', '  aA  ', 'abBABba'], dtype='<U7')
-    >>> (np.char.lstrip(c, ' ') == np.char.lstrip(c, '')).all()
-    np.False_
-    >>> (np.char.lstrip(c, ' ') == np.char.lstrip(c)).all()
-    np.True_
-
-    """
-    if chars is None:
-        return numpy._core.umath._lstrip_whitespace(a)
-    return numpy._core.umath._lstrip_chars(a, chars)
-
-
-def _partition_dispatcher(a, sep):
-    return (a,)
-
-
-@array_function_dispatch(_partition_dispatcher)
-def partition(a, sep):
-    """
-    Partition each element in `a` around `sep`.
-
-    Calls :meth:`str.partition` element-wise.
-
-    For each element in `a`, split the element as the first
-    occurrence of `sep`, and return 3 strings containing the part
-    before the separator, the separator itself, and the part after
-    the separator. If the separator is not found, return 3 strings
-    containing the string itself, followed by two empty strings.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array
-    sep : {str, unicode}
-        Separator to split each string element in `a`.
-
-    Returns
-    -------
-    out : ndarray, {str, unicode}
-        Output array of str or unicode, depending on input type.
-        The output array will have an extra dimension with 3
-        elements per input element.
-
-    Examples
-    --------
-    >>> x = np.array(["Numpy is nice!"])
-    >>> np.char.partition(x, " ")
-    array([['Numpy', ' ', 'is nice!']], dtype='<U8')
-    
-    See Also
-    --------
-    str.partition
-
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, 'partition', (sep,)), a)
-
-
-def _replace_dispatcher(a, old, new, count=None):
-    return (a,)
-
-
-@array_function_dispatch(_replace_dispatcher)
-def replace(a, old, new, count=None):
-    """
-    For each element in `a`, return a copy of the string with all
-    occurrences of substring `old` replaced by `new`.
-
-    Calls :meth:`str.replace` element-wise.
-
-    Parameters
-    ----------
-    a : array-like of str or unicode
-
-    old, new : scalar or array-like str or unicode
-
-    count : scalar or array-like int
-        If the optional argument `count` is given, only the first
-        `count` occurrences are replaced. If negative, replace all.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type.
-
-    See Also
-    --------
-    str.replace
-    
-    Examples
-    --------
-    >>> a = np.array(["That is a mango", "Monkeys eat mangos"])
-    >>> np.char.replace(a, 'mango', 'banana')
-    array(['That is a banana', 'Monkeys eat bananas'], dtype='<U19')
-
-    >>> a = np.array(["The dish is fresh", "This is it"])
-    >>> np.char.replace(a, 'is', 'was')
-    array(['The dwash was fresh', 'Thwas was it'], dtype='<U19')
-    
-    """
-    a_arr = numpy.asanyarray(a)
-    old = numpy.asanyarray(old)
-    new = numpy.asanyarray(new)
-    max_int64 = numpy.iinfo(numpy.int64).max
-    counts = numpy._core.umath.count(a_arr, old, 0, max_int64)
-    if count is not None:
-        count = numpy.asanyarray(count)
-        counts = numpy.where(count < 0, counts,
-                             numpy.minimum(counts, count))
-
-    buffersizes = (
-        numpy._core.umath.str_len(a_arr)
-        + counts * (numpy._core.umath.str_len(new) -
-                    numpy._core.umath.str_len(old))
-    )
-    # buffersizes is properly broadcast along all inputs.
-    out = numpy.empty_like(a_arr, shape=buffersizes.shape,
-                           dtype=f"{a_arr.dtype.char}{buffersizes.max()}")
-    numpy._core.umath._replace(a_arr, old, new, counts, out=out)
-    return out
-
-
-@array_function_dispatch(_count_dispatcher)
-def rfind(a, sub, start=0, end=None):
-    """
-    For each element in `a`, return the highest index in the string
-    where substring `sub` is found, such that `sub` is contained
-    within [`start`, `end`].
-
-    Calls :meth:`str.rfind` element-wise.
-
-    Parameters
-    ----------
-    a : array-like of str or unicode
-
-    sub : str or unicode
-
-    start, end : int, optional
-        Optional arguments `start` and `end` are interpreted as in
-        slice notation.
-
-    Returns
-    -------
-    out : ndarray
-       Output array of ints.  Return -1 on failure.
-
-    See Also
-    --------
-    str.rfind
-
-    Examples
-    --------
-    >>> a = np.array(["NumPy is a Python library"])
-    >>> np.char.rfind(a, "Python", start=0, end=None)
-    array([11])
-    
-    """
-    end = end if end is not None else numpy.iinfo(numpy.int64).max
-    return numpy._core.umath.rfind(a, sub, start, end)
-
-
-@array_function_dispatch(_count_dispatcher)
-def rindex(a, sub, start=0, end=None):
-    """
-    Like `rfind`, but raises :exc:`ValueError` when the substring `sub` is
-    not found.
-
-    Calls :meth:`str.rindex` element-wise.
-
-    Parameters
-    ----------
-    a : array-like of str or unicode
-
-    sub : str or unicode
-
-    start, end : int, optional
-
-    Returns
-    -------
-    out : ndarray
-       Output array of ints.
-
-    See Also
-    --------
-    rfind, str.rindex
-
-    Examples
-    --------
-    >>> a = np.array(["Computer Science"])
-    >>> np.char.rindex(a, "Science", start=0, end=None)
-    array([9])
-    
-    """
-    return _vec_string(
-        a, int_, 'rindex', [sub, start] + _clean_args(end))
-
-
-@array_function_dispatch(_just_dispatcher)
-def rjust(a, width, fillchar=' '):
-    """
-    Return an array with the elements of `a` right-justified in a
-    string of length `width`.
-
-    Calls :meth:`str.rjust` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    width : int
-        The length of the resulting strings
-    fillchar : str or unicode, optional
-        The character to use for padding
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.rjust
-
-    Examples
-    --------
-    >>> a = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> np.char.rjust(a, width=3)
-    array(['aAa', '  a', 'abB'], dtype='<U3')
-    
-    """
-    a_arr = numpy.asarray(a)
-    width_arr = numpy.asarray(width)
-    size = int(numpy.max(width_arr.flat))
-    if numpy.issubdtype(a_arr.dtype, numpy.bytes_):
-        fillchar = asbytes(fillchar)
-    return _vec_string(
-        a_arr, type(a_arr.dtype)(size), 'rjust', (width_arr, fillchar))
-
-
-@array_function_dispatch(_partition_dispatcher)
-def rpartition(a, sep):
-    """
-    Partition (split) each element around the right-most separator.
-
-    Calls :meth:`str.rpartition` element-wise.
-
-    For each element in `a`, split the element as the last
-    occurrence of `sep`, and return 3 strings containing the part
-    before the separator, the separator itself, and the part after
-    the separator. If the separator is not found, return 3 strings
-    containing the string itself, followed by two empty strings.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-        Input array
-    sep : str or unicode
-        Right-most separator to split each element in array.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of string or unicode, depending on input
-        type.  The output array will have an extra dimension with
-        3 elements per input element.
-
-    See Also
-    --------
-    str.rpartition
-
-    Examples
-    --------
-    >>> a = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> np.char.rpartition(a, 'A')
-    array([['aAaAa', 'A', ''],
-       ['  a', 'A', '  '],
-       ['abB', 'A', 'Bba']], dtype='<U5')
-
-    """
-    return _to_bytes_or_str_array(
-        _vec_string(a, object_, 'rpartition', (sep,)), a)
-
-
-def _split_dispatcher(a, sep=None, maxsplit=None):
-    return (a,)
-
-
-@array_function_dispatch(_split_dispatcher)
-def rsplit(a, sep=None, maxsplit=None):
-    """
-    For each element in `a`, return a list of the words in the
-    string, using `sep` as the delimiter string.
-
-    Calls :meth:`str.rsplit` element-wise.
-
-    Except for splitting from the right, `rsplit`
-    behaves like `split`.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    sep : str or unicode, optional
-        If `sep` is not specified or None, any whitespace string
-        is a separator.
-    maxsplit : int, optional
-        If `maxsplit` is given, at most `maxsplit` splits are done,
-        the rightmost ones.
-
-    Returns
-    -------
-    out : ndarray
-       Array of list objects
-
-    See Also
-    --------
-    str.rsplit, split
-
-    Examples
-    --------
-    >>> a = np.array(['aAaAaA', 'abBABba'])
-    >>> np.char.rsplit(a, 'A')
-    array([list(['a', 'a', 'a', '']), list(['abB', 'Bba'])], dtype=object)
-    
-    """
-    # This will return an array of lists of different sizes, so we
-    # leave it as an object array
-    return _vec_string(
-        a, object_, 'rsplit', [sep] + _clean_args(maxsplit))
-
-
-def _strip_dispatcher(a, chars=None):
-    return (a,)
-
-
-@array_function_dispatch(_strip_dispatcher)
-def rstrip(a, chars=None):
-    """
-    For each element in `a`, return a copy with the trailing characters
-    removed.
-
-    Parameters
-    ----------
-    a : array_like, with ``bytes_`` or ``unicode_`` dtype
-    chars : scalar with the same dtype as ``a``, optional
-       The ``chars`` argument is a string specifying the set of
-       characters to be removed. If ``None``, the ``chars``
-       argument defaults to removing whitespace. The ``chars`` argument
-       is not a prefix or suffix; rather, all combinations of its
-       values are stripped.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of ``bytes_`` or ``unicode_`` dtype
-
-    See Also
-    --------
-    str.rstrip
-
-    Examples
-    --------
-    >>> c = np.array(['aAaAaA', 'abBABba'])
-    >>> c
-    array(['aAaAaA', 'abBABba'], dtype='<U7')
-    >>> np.char.rstrip(c, 'a')
-    array(['aAaAaA', 'abBABb'], dtype='<U7')
-    >>> np.char.rstrip(c, 'A')
-    array(['aAaAa', 'abBABba'], dtype='<U7')
-
-    """
-    if chars is None:
-        return numpy._core.umath._rstrip_whitespace(a)
-    return numpy._core.umath._rstrip_chars(a, chars)
-
-
-@array_function_dispatch(_split_dispatcher)
-def split(a, sep=None, maxsplit=None):
-    """
-    For each element in `a`, return a list of the words in the
-    string, using `sep` as the delimiter string.
-
-    Calls :meth:`str.split` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    sep : str or unicode, optional
-       If `sep` is not specified or None, any whitespace string is a
-       separator.
-
-    maxsplit : int, optional
-        If `maxsplit` is given, at most `maxsplit` splits are done.
-
-    Returns
-    -------
-    out : ndarray
-        Array of list objects
-
-    Examples
-    --------
-    >>> x = np.array("Numpy is nice!")
-    >>> np.char.split(x, " ")
-    array(list(['Numpy', 'is', 'nice!']), dtype=object)
-
-    >>> np.char.split(x, " ", 1)
-    array(list(['Numpy', 'is nice!']), dtype=object)
-
-    See Also
-    --------
-    str.split, rsplit
-
-    """
-    # This will return an array of lists of different sizes, so we
-    # leave it as an object array
-    return _vec_string(
-        a, object_, 'split', [sep] + _clean_args(maxsplit))
-
-
-def _splitlines_dispatcher(a, keepends=None):
-    return (a,)
-
-
-@array_function_dispatch(_splitlines_dispatcher)
-def splitlines(a, keepends=None):
-    """
-    For each element in `a`, return a list of the lines in the
-    element, breaking at line boundaries.
-
-    Calls :meth:`str.splitlines` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    keepends : bool, optional
-        Line breaks are not included in the resulting list unless
-        keepends is given and true.
-
-    Returns
-    -------
-    out : ndarray
-        Array of list objects
-
-    See Also
-    --------
-    str.splitlines
-
-    """
-    return _vec_string(
-        a, object_, 'splitlines', _clean_args(keepends))
-
-
-def _startswith_dispatcher(a, prefix, start=None, end=None):
-    return (a,)
-
-
-@array_function_dispatch(_startswith_dispatcher)
-def startswith(a, prefix, start=0, end=None):
-    """
-    Returns a boolean array which is `True` where the string element
-    in `a` starts with `prefix`, otherwise `False`.
-
-    Calls :meth:`str.startswith` element-wise.
-
-    Parameters
-    ----------
-    a : array_like of str or unicode
-
-    prefix : str
-
-    start, end : int, optional
-        With optional `start`, test beginning at that position. With
-        optional `end`, stop comparing at that position.
-
-    Returns
-    -------
-    out : ndarray
-        Array of booleans
-
-    See Also
-    --------
-    str.startswith
-
-    """
-    end = end if end is not None else numpy.iinfo(numpy.int_).max
-    return numpy._core.umath.startswith(a, prefix, start, end)
-
-
-@array_function_dispatch(_strip_dispatcher)
-def strip(a, chars=None):
-    """
-    For each element in `a`, return a copy with the leading and
-    trailing characters removed.
-
-    Parameters
-    ----------
-    a : array_like, with ``bytes_`` or ``unicode_`` dtype
-    chars : scalar with the same dtype as ``a``, optional
-       The ``chars`` argument is a string specifying the set of
-       characters to be removed. If None, the ``chars``
-       argument defaults to removing whitespace. The ``chars`` argument
-       is not a prefix or suffix; rather, all combinations of its
-       values are stripped.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of ``bytes_`` or ``unicode_`` dtype
-
-    See Also
-    --------
-    str.strip
-
-    Examples
-    --------
-    >>> c = np.array(['aAaAaA', '  aA  ', 'abBABba'])
-    >>> c
-    array(['aAaAaA', '  aA  ', 'abBABba'], dtype='<U7')
-    >>> np.char.strip(c)
-    array(['aAaAaA', 'aA', 'abBABba'], dtype='<U7')
-    # 'a' unstripped from c[1] because of leading whitespace.
-    >>> np.char.strip(c, 'a')
-    array(['AaAaA', '  aA  ', 'bBABb'], dtype='<U7')
-    # 'A' unstripped from c[1] because of trailing whitespace.
-    >>> np.char.strip(c, 'A')
-    array(['aAaAa', '  aA  ', 'abBABba'], dtype='<U7')
-
-    """
-    if chars is None:
-        return numpy._core.umath._strip_whitespace(a)
-    return numpy._core.umath._strip_chars(a, chars)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def swapcase(a):
-    """
-    Return element-wise a copy of the string with
-    uppercase characters converted to lowercase and vice versa.
-
-    Calls :meth:`str.swapcase` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array.
-
-    Returns
-    -------
-    out : ndarray, {str, unicode}
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.swapcase
-
-    Examples
-    --------
-    >>> c=np.array(['a1B c','1b Ca','b Ca1','cA1b'],'S5'); c
-    array(['a1B c', '1b Ca', 'b Ca1', 'cA1b'],
-        dtype='|S5')
-    >>> np.char.swapcase(c)
-    array(['A1b C', '1B cA', 'B cA1', 'Ca1B'],
-        dtype='|S5')
-
-    """
-    a_arr = numpy.asarray(a)
-    return _vec_string(a_arr, a_arr.dtype, 'swapcase')
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def title(a):
-    """
-    Return element-wise title cased version of string or unicode.
-
-    Title case words start with uppercase characters, all remaining cased
-    characters are lowercase.
-
-    Calls :meth:`str.title` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array.
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.title
-
-    Examples
-    --------
-    >>> c=np.array(['a1b c','1b ca','b ca1','ca1b'],'S5'); c
-    array(['a1b c', '1b ca', 'b ca1', 'ca1b'],
-        dtype='|S5')
-    >>> np.char.title(c)
-    array(['A1B C', '1B Ca', 'B Ca1', 'Ca1B'],
-        dtype='|S5')
-
-    """
-    a_arr = numpy.asarray(a)
-    return _vec_string(a_arr, a_arr.dtype, 'title')
-
-
-def _translate_dispatcher(a, table, deletechars=None):
-    return (a,)
-
-
-@array_function_dispatch(_translate_dispatcher)
-def translate(a, table, deletechars=None):
-    """
-    For each element in `a`, return a copy of the string where all
-    characters occurring in the optional argument `deletechars` are
-    removed, and the remaining characters have been mapped through the
-    given translation table.
-
-    Calls :meth:`str.translate` element-wise.
-
-    Parameters
-    ----------
-    a : array-like of str or unicode
-
-    table : str of length 256
-
-    deletechars : str
-
-    Returns
-    -------
-    out : ndarray
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.translate
-
-    Examples
-    --------
-    >>> a = np.array(['a1b c', '1bca', 'bca1'])
-    >>> table = a[0].maketrans('abc', '123')
-    >>> deletechars = ' '
-    >>> np.char.translate(a, table, deletechars)
-    array(['112 3', '1231', '2311'], dtype='<U5')
-    
-    """
-    a_arr = numpy.asarray(a)
-    if issubclass(a_arr.dtype.type, str_):
-        return _vec_string(
-            a_arr, a_arr.dtype, 'translate', (table,))
-    else:
-        return _vec_string(
-            a_arr,
-            a_arr.dtype,
-            'translate',
-            [table] + _clean_args(deletechars)
-        )
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def upper(a):
-    """
-    Return an array with the elements converted to uppercase.
-
-    Calls :meth:`str.upper` element-wise.
-
-    For 8-bit strings, this method is locale-dependent.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array.
-
-    Returns
-    -------
-    out : ndarray, {str, unicode}
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.upper
-
-    Examples
-    --------
-    >>> c = np.array(['a1b c', '1bca', 'bca1']); c
-    array(['a1b c', '1bca', 'bca1'], dtype='<U5')
-    >>> np.char.upper(c)
-    array(['A1B C', '1BCA', 'BCA1'], dtype='<U5')
-
-    """
-    a_arr = numpy.asarray(a)
-    return _vec_string(a_arr, a_arr.dtype, 'upper')
-
-
-def _zfill_dispatcher(a, width):
-    return (a,)
-
-
-@array_function_dispatch(_zfill_dispatcher)
-def zfill(a, width):
-    """
-    Return the numeric string left-filled with zeros
-
-    Calls :meth:`str.zfill` element-wise.
-
-    Parameters
-    ----------
-    a : array_like, {str, unicode}
-        Input array.
-    width : int
-        Width of string to left-fill elements in `a`.
-
-    Returns
-    -------
-    out : ndarray, {str, unicode}
-        Output array of str or unicode, depending on input type
-
-    See Also
-    --------
-    str.zfill
-
-    Examples
-    --------
-    >>> np.char.zfill('1', 3)
-    array('001', dtype='<U3')
-
-    """
-    a_arr = numpy.asarray(a)
-    width_arr = numpy.asarray(width)
-    size = int(numpy.max(width_arr.flat))
-    return _vec_string(
-        a_arr, type(a_arr.dtype)(size), 'zfill', (width_arr,))
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isnumeric(a):
-    """
-    For each element, return True if there are only numeric
-    characters in the element.
-
-    Calls :meth:`str.isnumeric` element-wise.
-
-    Numeric characters include digit characters, and all characters
-    that have the Unicode numeric value property, e.g. ``U+2155,
-    VULGAR FRACTION ONE FIFTH``.
-
-    Parameters
-    ----------
-    a : array_like, unicode
-        Input array.
-
-    Returns
-    -------
-    out : ndarray, bool
-        Array of booleans of same shape as `a`.
-
-    See Also
-    --------
-    str.isnumeric
-
-    Examples
-    --------
-    >>> np.char.isnumeric(['123', '123abc', '9.0', '1/4', 'VIII'])
-    array([ True, False, False, False, False])
-
-    """
-    return numpy._core.umath.isnumeric(a)
-
-
-@array_function_dispatch(_unary_op_dispatcher)
-def isdecimal(a):
-    """
-    For each element, return True if there are only decimal
-    characters in the element.
-
-    Calls :meth:`str.isdecimal` element-wise.
-
-    Decimal characters include digit characters, and all characters
-    that can be used to form decimal-radix numbers,
-    e.g. ``U+0660, ARABIC-INDIC DIGIT ZERO``.
-
-    Parameters
-    ----------
-    a : array_like, unicode
-        Input array.
-
-    Returns
-    -------
-    out : ndarray, bool
-        Array of booleans identical in shape to `a`.
-
-    See Also
-    --------
-    str.isdecimal
-
-    Examples
-    --------
-    >>> np.char.isdecimal(['12345', '4.99', '123ABC', ''])
-    array([ True, False, False, False])
-
-    """ 
-    return numpy._core.umath.isdecimal(a)
 
 
 @set_module("numpy.char")
@@ -2690,7 +840,7 @@ class chararray(ndarray):
         """
         return asarray(partition(self, sep))
 
-    def replace(self, old, new, count=None):
+    def replace(self, old, new, count=-1):
         """
         For each element in `self`, return a copy of the string with all
         occurrences of substring `old` replaced by `new`.
@@ -2987,7 +1137,7 @@ def array(obj, itemsize=None, copy=True, unicode=None, order=None):
                          buffer=obj, order=order)
 
     if isinstance(obj, (list, tuple)):
-        obj = numpy.asarray(obj)
+        obj = asnarray(obj)
 
     if isinstance(obj, ndarray) and issubclass(obj.dtype.type, character):
         # If we just have a vanilla chararray, create a chararray
@@ -3015,7 +1165,7 @@ def array(obj, itemsize=None, copy=True, unicode=None, order=None):
             dtype = bytes_
 
         if order is not None:
-            obj = numpy.asarray(obj, order=order)
+            obj = asnarray(obj, order=order)
         if (copy or
                 (itemsize != obj.itemsize) or
                 (not unicode and isinstance(obj, str_)) or
