+++ numpy/strings/__init__.pyi
@@ -5,22 +5,49 @@ from numpy._core.strings import (
     less_equal as less_equal,
     greater as greater,
     less as less,
-    str_len as str_len,
     add as add,
-    count as count,
-    endswith as endswith,
-    find as find,
+    multiply as multiply,
+    mod as mod,
     isalpha as isalpha,
+    isalnum as isalnum,
     isdigit as isdigit,
     isspace as isspace,
-    lstrip as lstrip,
-    rfind as rfind,
-    rstrip as rstrip,
-    startswith as startswith,
-    strip as strip,
     isnumeric as isnumeric,
     isdecimal as isdecimal,
+    islower as islower,
+    isupper as isupper,
+    istitle as istitle,
+    str_len as str_len,
+    find as find,
+    rfind as rfind,
+    index as index,
+    rindex as rindex,
+    count as count,
+    startswith as startswith,
+    endswith as endswith,
+    decode as decode,
+    encode as encode,
+    expandtabs as expandtabs,
+    center as center,
+    ljust as ljust,
+    rjust as rjust,
+    lstrip as lstrip,
+    rstrip as rstrip,
+    strip as strip,
+    zfill as zfill,
+    upper as upper,
+    lower as lower,
+    swapcase as swapcase,
+    capitalize as capitalize,
+    title as title,
     replace as replace,
+    join as join,
+    split as split,
+    rsplit as rsplit,
+    splitlines as splitlines,
+    partition as partition,
+    rpartition as rpartition,
+    translate as translate,
 )
 
 __all__: list[str]
